+++ cdc-office-tables/src/main/java/cdc/office/tables/Row.java
@@ -49,11 +49,6 @@ public interface Row {
     /**
      * @return The number of columns in the row.
      */
-    @Deprecated
-    public default int getColumnsCount() {
-        return size();
-    }
-
     public default int size() {
         return getValues().size();
     }